Supporting all the flipping/rotating of tiles.
TilesetTool now puts the pivot point in the center. This required shifting tile positions
and fixing how collisions are generated since they were off by half a unit.

However this made it much simpler to support the various flips and rotations by not having
to also shift the position of the tiles.